backend/requirements.txt
# Core frameworks
fastapi>=0.104.1
uvicorn>=0.23.2
starlette>=0.27.0
httpx>=0.25.1

# Database
sqlalchemy>=2.0.0
asyncpg>=0.27.0

# Authentication
python-jose>=3.3.0
passlib>=1.7.4

# Pydantic
pydantic>=2.4.2
pydantic-settings>=2.0.3

# LangChain ecosystem
langchain>=0.0.267
langchain-community>=0.0.10
langchain-huggingface>=0.0.8
langchain-openai>=0.0.2
huggingface-hub>=0.16.4

# Document processing
PyPDF2>=3.0.0
python-docx>=0.8.11

# Vector stores and embeddings
faiss-cpu>=1.7.4
sentence-transformers==2.2.2
chromadb>=0.4.6
torch==2.0.1
transformers==4.30.2
numpy==1.24.3

# Utilities
python-dotenv>=1.0.0
websockets>=11.0.3
jsonschema>=4.19.1
requests>=2.28.0

# Time handling
datetime

# Form data
python-multipart>=0.0.5

# Email validator
email-validator>=2.0.0